Refactor of client code
- Pylint fixes
- Better logging
- Split functionality a bit more

+++ backup_tool/client.py
@@ -1,6 +1,7 @@
 import logging
 from logging.handlers import RotatingFileHandler
 import os
+import re
 import uuid
 
 from sqlalchemy import create_engine
@@ -85,48 +86,58 @@ class BackupClient():
             return False
 
         if not local_file.backup_entry_id:
-            self.logger.error("No backup entry for local file:%s" % local_file_id)
+            self.logger.error("No backup entry for local file:%s", local_file_id)
             return False
 
         backup_entry = self.db_session.query(BackupEntry).get(local_file.backup_entry_id)
+
+        if not backup_entry:
+            self.logger.error("Expecting backup entry %s does not exist", local_file.backup_entry_id)
 
         local_file_path = local_file.local_file_path
         if self.relative_path:
             local_file_path = os.path.join(self.relative_path, local_file_path)
 
         if os.path.isfile(local_file_path):
+            self.logger.debug("Checking local file %s md5", local_file_path)
             local_file_md5 = utils.md5(local_file_path)
+            self.logger.debug("Local file %s has md5 sum %s", local_file_path, local_file_md5)
             if local_file.local_md5_checksum == local_file_md5:
                 if not overwrite:
-                    self.logger.info("Local file md5 %s matches expected md5", local_file_md5)
+                    self.logger.info("Local file %s has expected md5 %s", local_file_path, local_file_md5)
                     return True
 
         # Write file to temp dir
         with utils.temp_file() as encrypted_file:
-            self.logger.debug("Using file %s for download", encrypted_file)
+            self.logger.info("Downloading object %s to temp file %s", backup_entry.uploaded_file_path, encrypted_file)
             self.os_client.object_get(self.oci_namespace, self.oci_bucket,
                                       backup_entry.uploaded_file_path, encrypted_file)
-            self.logger.info("Downloaded object %s to temp file %s", backup_entry.uploaded_file_path, encrypted_file)
+            self.logger.info("Downloaded of object %s complete, written to temp file %s", backup_entry.uploaded_file_path, encrypted_file)
 
             # Check md5 matches expected
+            self.logger.debug("Checking md5 sum of temp file %s", encrypted_file)
             downloaded_md5 = utils.md5(encrypted_file)
+            self.logger.debug("Downloaded encrypted file %s has md5 sum %s", encrypted_file, downloaded_md5)
             if backup_entry.uploaded_md5_checksum != downloaded_md5:
                 self.logger.error("Downloaded file %s has unexpected md5 %s, expected %s",
                                   encrypted_file, downloaded_md5, backup_entry.uploaded_md5_checksum)
                 return True
-            self.logger.debug("Decrypting file %s to file %s", encrypted_file, local_file_path)
+            self.logger.debug("Decrypting temp file %s to file %s", encrypted_file, local_file_path)
             dir_name = os.path.dirname(local_file_path)
             if not os.path.isdir(dir_name):
                 os.makedirs(dir_name)
             crypto.decrypt_file(encrypted_file, local_file_path, self.crypto_key,
                                 backup_entry.uploaded_encryption_offset)
 
-            # Check md5 matches expected
-            local_file_md5 = utils.md5(local_file_path)
-            if local_file_md5 != local_file.local_md5_checksum:
-                self.logger.error("MD5 %s of decrypted file %s does not match expected %s",
-                                  local_file_md5, local_file_path, local_file.local_md5_checksum)
-                return False
+
+        # Check md5 matches expected
+        self.logger.debug("Checking md5 of file %s", local_file_path)
+        local_file_md5 = utils.md5(local_file_path)
+        self.logger.debug("Local file %s has md5 sum %s", local_file_path, local_file_md5)
+        if local_file_md5 != local_file.local_md5_checksum:
+            self.logger.error("MD5 %s of decrypted file %s does not match expected %s",
+                              local_file_md5, local_file_path, local_file.local_md5_checksum)
+            return False
         return True
 
 
@@ -144,9 +155,90 @@ class BackupClient():
 
             local_file          :       Full path of local file
             overwrite           :       Upload new file is md5 is changed
-            check_uploaded_md5  :       Ensure any existing backup file matches expected encryption 
+            check_uploaded_md5  :       Ensure any existing backup file matches expected encryption
         '''
         self._file_backup(local_file, overwrite=overwrite, check_uploaded_md5=check_uploaded_md5)
+
+    def _file_backup_file_exists(self, local_backup_file, local_file, local_file_md5, overwrite, check_uploaded_md5):
+        '''
+            Local backup of file exists
+        '''
+        # Keep boolean value to make sure we should upload new file
+        upload_file = True
+
+        self.logger.debug("Found existing local file:%s", local_backup_file.id)
+        if local_file_md5 == local_backup_file.local_md5_checksum:
+            self.logger.debug("Existing local file %s has expected md5 %s", local_file, local_file_md5)
+            # Only upload if no backup file exists
+            # If requested, check that backup file matches encryption
+            if local_backup_file.backup_entry_id is not None and check_uploaded_md5 is False:
+                upload_file = False
+        else:
+            self.logger.debug("Existing local file %s has unexpected  md5 sum %s", local_file, local_file_md5)
+            if overwrite is True:
+                local_backup_file.local_md5_checksum = local_file_md5
+                # Current file has no backup, so set this to null for now
+                local_backup_file.backup_entry_id = None
+                self.db_session.commit()
+                self.logger.debug("Updated local file %s to checksum %s",
+                                  local_backup_file.id, local_file_md5)
+            else:
+                self.logger.warning("Overwrite set to false, not uploading new version of local file %s", local_file)
+                upload_file = False
+        return upload_file
+
+    def _file_backup_upload(self, local_file, local_backup_file):
+        with utils.temp_file() as crypto_file:
+            self.logger.debug("Creating encrypted file %s from file %s with md5",
+                              crypto_file, local_file)
+            offset = crypto.encrypt_file(local_file, crypto_file, self.crypto_key)
+            self.logger.info("Created encrypted file %s from file %s with offset %s",
+                             crypto_file, local_file, offset)
+            self.logger.debug("Checking md5 sum of file %s", crypto_file)
+            local_crypto_file_md5 = utils.md5(crypto_file)
+            self.logger.info("Encrypted file %s has md5 sum %s", crypto_file, local_crypto_file_md5)
+
+            # Check if md5 file already exists
+            backup_entry = self.db_session.query(BackupEntry).\
+                    filter(BackupEntry.uploaded_md5_checksum == local_crypto_file_md5).first()
+
+            if backup_entry:
+                # If file exists, just upload local files data
+                self.logger.debug("Found existing upload with matching md5 found %s for file %s",
+                                  backup_entry.id, local_file)
+                local_backup_file.backup_entry_id = backup_entry.id
+                self.db_session.commit()
+                self.logger.info("Updated local file %s with backup entry %s", local_backup_file.id, backup_entry.id)
+                return True
+
+            # Else upload new file
+            self.logger.info("No encrypted upload matching file with md5 %s, uploading copy",
+                             local_crypto_file_md5)
+
+            object_path = self._generate_uuid()
+
+            self.logger.debug("Uploading encrypted file %s to object path %s",
+                              crypto_file, object_path)
+            self.os_client.object_put(self.oci_namespace, self.oci_bucket, object_path, crypto_file,
+                                      md5_sum=local_crypto_file_md5)
+
+        backup_args = {
+            'uploaded_file_path' : object_path,
+            'uploaded_md5_checksum' : local_crypto_file_md5,
+            'uploaded_encryption_offset' : offset,
+        }
+
+        backup_entry = BackupEntry(**backup_args)
+        self.db_session.add(backup_entry)
+        self.db_session.commit()
+        self.logger.info("Uploaded encrypted file %s as backup entry %s",
+                         crypto_file, backup_entry.id)
+
+        local_backup_file.backup_entry_id = backup_entry.id
+        self.db_session.commit()
+        self.logger.info("Updated local backup %s to match backup entry %s",
+                         local_backup_file.id, backup_entry.id)
+        return True
 
     def _file_backup(self, local_file, overwrite=True, check_uploaded_md5=False):
         '''
@@ -154,7 +246,7 @@ class BackupClient():
 
             local_file          :       Full path of local file
             overwrite           :       Upload new file is md5 is changed
-            check_uploaded_md5  :       Ensure any existing backup file matches expected encryption 
+            check_uploaded_md5  :       Ensure any existing backup file matches expected encryption
         '''
         # Use local file as the full path of the file
         # Use local file path as relative path for the database
@@ -167,36 +259,19 @@ class BackupClient():
         if local_file_path != local_file:
             self.logger.debug("Using relative path for database %s", local_file_path)
 
-        # Keep boolean value to make sure we should upload new file
-        upload_file = True
 
         # First check if local file exists, and if so, has md5 changed
+        self.logger.debug("Checking md5 sum for local file %s", local_file)
         local_file_md5 = utils.md5(local_file)
         self.logger.debug("Local file %s md5 sum:%s", local_file, local_file_md5)
 
         local_backup_file = self.db_session.query(BackupEntryLocalFile).\
                 filter(BackupEntryLocalFile.local_file_path == local_file_path).first()
         if local_backup_file:
-            self.logger.debug("Found existing local file:%s", local_backup_file.id)
-            if local_file_md5 == local_backup_file.local_md5_checksum:
-                self.logger.debug("Existing local file has same md5")
-                # Only upload if no backup file exists
-                # If requested, check that backup file matches encryption
-                if local_backup_file.backup_entry_id is not None and check_uploaded_md5 is False:
-                    upload_file = False
-            else:
-                self.logger.debug("Existing local file has different md5 sum, updating")
-                if overwrite is True:
-                    local_backup_file.local_md5_checksum = local_file_md5
-                    # Current file has no backup, so set this to null for now
-                    local_backup_file.backup_entry_id = None
-                    self.db_session.commit()
-                    self.logger.debug("Updated local file %s to checksum %s",
-                                      local_backup_file.id, local_file_md5)
-                else:
-                    upload_file = False
-
+            upload_file = self._file_backup_file_exists(local_backup_file, local_file,
+                                                        local_file_md5, overwrite, check_uploaded_md5)
         else:
+            upload_file = True
             self.logger.debug("No existing local file found for path:%s", local_file)
             backup_file_args = {
                 'local_file_path': local_file_path,
@@ -206,53 +281,11 @@ class BackupClient():
             local_backup_file = BackupEntryLocalFile(**backup_file_args)
             self.db_session.add(local_backup_file)
             self.db_session.commit()
+            self.logger.info("Created database entry %s for local file %s", local_backup_file.id, local_file)
 
         if upload_file:
-            # Encrypt local file
-            with utils.temp_file() as crypto_file:
-                offset = crypto.encrypt_file(local_file, crypto_file, self.crypto_key)
-                local_crypto_file_md5 = utils.md5(crypto_file)
-                self.logger.debug("Created encrypted file %s from file %s with md5 %s",
-                                  crypto_file, local_file, local_crypto_file_md5)
-
-                # Check if md5 file already exists
-                backup_entry = self.db_session.query(BackupEntry).\
-                        filter(BackupEntry.uploaded_md5_checksum == local_crypto_file_md5).first()
-
-                if backup_entry:
-                    # If file exists, just upload local files data
-                    self.logger.info("Encrypted upload with matching md5 found %s for file %s",
-                                     backup_entry.id, local_file)
-                    local_backup_file.backup_entry_id = backup_entry.id
-                    self.db_session.commit()
-                else:
-                    # Else upload new file
-                    self.logger.info("No encrypted upload matching file %s, uploading copy",
-                                     local_file)
-
-                    object_path = self._generate_uuid()
-
-                    self.logger.debug("Uploading encrypted file %s to object path %s",
-                                      crypto_file, object_path)
-                    self.os_client.object_put(self.oci_namespace, self.oci_bucket, object_path, crypto_file,
-                                              md5_sum=local_crypto_file_md5)
-
-                    backup_args = {
-                        'uploaded_file_path' : object_path,
-                        'uploaded_md5_checksum' : local_crypto_file_md5,
-                        'uploaded_encryption_offset' : offset,
-                    }
-
-                    backup_entry = BackupEntry(**backup_args)
-                    self.db_session.add(backup_entry)
-                    self.db_session.commit()
-                    self.logger.info("Uploaded encrypted file %s as backup entry %s",
-                                     crypto_file, backup_entry.id)
-
-                    local_backup_file.backup_entry_id = backup_entry.id
-                    self.db_session.commit()
-                    self.logger.info("Updated local backup %s to match backup entry %s",
-                                     local_backup_file.id, backup_entry.id)
+            return self._file_backup_upload(local_file, local_backup_file)
+        return True
 
     def file_duplicates(self):
         '''
@@ -298,18 +331,22 @@ class BackupClient():
             # Check if dir matches skip files
             for skip_check in skip_files:
                 if re.match(skip_check, dir_name):
-                    self.logger.warn("Ignoring dir %s since matches skip check %s", dir_name, skip_check)
+                    self.logger.warning("Ignoring dir %s since matches skip check %s", dir_name, skip_check)
                     continue
             self.logger.info("Backing up directory %s", dir_name)
             for file_name in file_list:
                 full_path = os.path.join(dir_name, file_name)
+                # Skip if matches any continue
+                skip = False
                 for skip_check in skip_files:
                     if re.match(skip_check, full_path):
-                        self.logger.warn("Ignoring file %s since matches skip check %s", full_path, skip_check)
-                        continue
-                self._file_backup(full_path,
-                                  overwrite=overwrite,
-                                  check_uploaded_md5=check_uploaded_md5)
+                        self.logger.warning("Ignoring file %s since matches skip check %s", full_path, skip_check)
+                        skip = True
+                        break
+                if not skip:
+                    self._file_backup(full_path,
+                                      overwrite=overwrite,
+                                      check_uploaded_md5=check_uploaded_md5)
 
     def file_list(self):
         '''
@@ -320,20 +357,28 @@ class BackupClient():
             local_files.append(local_file.as_dict())
         return local_files
 
-    def file_cleanup(self):
+    def file_cleanup(self, dry_run=False):
         '''
             Delete local files in database that are no longer present on file system
+
+            dry_run     :       Return list of files, but do not execute deletes
         '''
+
+        files_cleaned = []
+
         for local_file in self.db_session.query(BackupEntryLocalFile).all():
             local_file_path = local_file.local_file_path
             if self.relative_path:
                 local_file_path = os.path.join(self.relative_path, local_file_path)
             if not os.path.isfile(local_file_path):
                 self.logger.info("Local file %s path %s no longer present, removing from db", local_file.id, local_file_path)
-                self.db_session.query(BackupEntryLocalFile).filter_by(id=local_file.id).delete()
-                self.db_session.commit()
+                if not dry_run:
+                    self.db_session.query(BackupEntryLocalFile).filter_by(id=local_file.id).delete()
+                    self.db_session.commit()
+                files_cleaned.append(local_file.id)
             else:
                 self.logger.debug("Local file %s path %s exists, skipping", local_file.id, local_file_path)
+        return files_cleaned
 
 
     def backup_list(self):
